fix_bugs
Merge all child components in parent component and send data with props to child component
